git-gui: warn if the commit message contains lines longer than the set limit
The commit message widget does not wrap the text and has a configurable
fixed width to avoid creating too wide commit messages. Though this is
only enforced visually by eliding the too long lines the GUI. Now we also
check the commit message at commit time for long lines and ask the author
for confirmation if it exceeds the configured line length.

Needs Tcl 8.5 because of `mathfunc::max` and the `{*}` thingy.

Signed-off-by: Bert Wesarg <bert.wesarg@googlemail.com>